+++ test/index_test.js
@@ -153,6 +153,12 @@ describe("Range", function() {
     })
   })
 
+  describe(".prototype.isFinite", function() {
+    it("must be an alias to Range.prototype.isBounded", function() {
+      Range.prototype.isFinite.must.equal(Range.prototype.isBounded)
+    })
+  })
+
   describe(".prototype.contains", function() {
     it("must return true when contained", function() {
       new Range(10, 20).contains(15).must.be.true()